Student Management app

+++ Student Management System/Student.java
@@ -3,16 +3,16 @@ public class Student {
     private int id;
     private String name;
     private String course;
-    private int marks;
+    private double marks;
 
-    public Student(int id, String name, String course, int marks) {
+    public Student(int id, String name, String course, double marks) {
         this.id = id;
         this.name = name;
         this.course = course;
         this.marks = marks;
     }
 
-    public int getId(){
+    public double getId(){
         return id;
     }
     public String getName(){
@@ -34,7 +34,7 @@ public class Student {
     public void setCourse(String course){
         this.course=course;
     }
-    public void setMarks(int marks){
+    public void setMarks(double marks){
         this.marks=marks;
     }
 }
